🔧 fix(dev-server): Update Ansible host IP address Changed the Ansible host IP from 54.157.240.227 to 3.239.63.193 to reflect the new server configuration.
🔧 fix(install-docker): Specify 'dev' as the target host
Updated the target hosts in the Ansible playbook from 'all' to 'dev' to ensure the playbook runs only on the development server.

🔧 fix(Jenkinsfile): Update DockerHub ID and server configuration
Changed the DockerHub ID from 'ulrichsteve' to 'kilann31' and updated the HOST_IP to 3.239.63.193. Also, enclosed APP_EXPOSED and CONTAINER_PORT values in quotes for consistency. These changes ensure the Jenkins pipeline uses the correct DockerHub account and server settings.